Move discovery stuff to separate file
